+++ app.js
@@ -2,6 +2,7 @@ const express = require('express');
 const morgan = require('morgan');
 const dotenv = require('dotenv');
 const diary_router = require('./routers/diary_router');
+const page_router = require('./routers/page_router');
 const auth_router = require('./routers/auth_router');
 const category_router = require('./routers/category_router');
 const cors = require('cors');
@@ -24,6 +25,7 @@ app.get('/', (req, res) => {
 });
 
 app.use('/auth', auth_router);
+app.use('/diaries/:diary_id/pages', page_router);
 app.use('/diaries', diary_router);
 app.use('/category', category_router);
